Revert "feat: Update character feature concluded and refactor"
