+++ api/src/main/java/com/naruto/api/characters/Character.java
@@ -1,14 +1,11 @@
-package com.naruto.api.characters.repository;
+package com.naruto.api.characters;
 
 import com.google.gson.Gson;
-import com.naruto.api.characters.dto.CharacterPutDTO;
-import com.naruto.api.characters.dto.CharactersPostDTO;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 
 @Table(name = "characters")
 @Entity(name = "Character")
@@ -56,23 +53,4 @@ public class Character {
         this.tools = gson.toJson(data.tools());
     }
 
-    public void updateInfo(CharacterPutDTO data) {
-        Gson gson = new Gson();
-
-        if (data.name() != null) {
-            this.name = data.name();
-        }
-        if (data.jutsu() != null) {
-            this.jutsu = gson.toJson(data.jutsu());
-        }
-        if (data.natureType() != null) {
-            this.natureType = gson.toJson(data.natureType());
-        }
-        if (data.personal() != null) {
-            this.personal = new Personal(data.personal());
-        }
-        if (data.tools() != null) {
-            this.tools = gson.toJson(data.tools());
-        }
-    }
 }
