+++ api/src/main/java/com/naruto/api/characters/CharactersPostDTO.java
@@ -1,11 +1,12 @@
-package com.naruto.api.characters.dto;
+package com.naruto.api.characters;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record CharacterPutDTO(
+public record CharactersPostDTO(
         @NotNull
         Long id,
+        @NotBlank
         String name,
         String[] jutsu,
         String[] natureType,
