+++ api/src/main/resources/db/migration/V01__create-table-characters.sql
