--- naruto-api/src/main/java/com/naruto/api/characters/dto/CharactersPostDTO.java
@@ -1,15 +1,0 @@
-package com.naruto.api.characters.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record CharactersPostDTO(
-        @NotNull
-        Long id,
-        @NotBlank
-        String name,
-        String[] jutsu,
-        String[] natureType,
-        PersonalPostDTO personal,
-        String[] tools
-) {}
